requirements.txt
Django==4.2.9
djangorestframework==3.14.0
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
